Upgrading projects to .net6 Updating API of Bgfx to latest (e6125ffa462a759fe67c0d7d5d95d3b5cce313b6)

+++ SharpBgfx/IndirectBuffer.cs
@@ -16,8 +16,8 @@ namespace SharpBgfx {
         /// Initializes a new instance of the <see cref="IndirectBuffer"/> struct.
         /// </summary>
         /// <param name="size">The number of commands that can fit in the buffer.</param>
-        public IndirectBuffer (int size) {
-            handle = NativeMethods.bgfx_create_indirect_buffer(size);
+        public IndirectBuffer (uint size) {
+            handle = NativeMethods.bgfx_create_indirect_buffer((uint)size);
         }
 
         /// <summary>
